comments

+++ FinalProject/Controllers/CommentsController.cs
@@ -19,13 +19,27 @@ namespace FinalProject.Controllers
         // GET: Comments
         public ActionResult Index()
         {
-            string strCurrentUserId = User.Identity.GetUserId();
-            ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(strCurrentUserId);
+            if (User.IsInRole("Admin"))
+            {
+                var comments = db.Comments;
+                return View(comments.ToList());
+            }
+            else if (User.Identity.IsAuthenticated)
+            {
+                string strCurrentUserId = User.Identity.GetUserId();
+                ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(strCurrentUserId);
 
-            var userName = user.FirstName + " " + user.LastName;
-            
-            var comments = db.Comments.Include(c => c.ArticleComment).Where(c => c.CommentUser == userName);
-            return View(comments.ToList());
+                var userName = user.FirstName + " " + user.LastName;
+
+                var comments = db.Comments.Include(c => c.ArticleComment).Where(c => c.CommentUser == userName);
+                return View(comments.ToList());
+            }
+            else
+            {
+                return RedirectToAction("Login", "Account");
+            }
+
+
         }
 
         // GET: Comments/Details/5
@@ -40,7 +54,31 @@ namespace FinalProject.Controllers
             {
                 return HttpNotFound();
             }
-            return View(comment);
+
+            if (User.IsInRole("Admin"))
+            {
+                return View(comment);
+            }
+            else if (User.Identity.IsAuthenticated)
+            {
+                string strCurrentUserId = User.Identity.GetUserId();
+                ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(strCurrentUserId);
+
+                var userName = user.FirstName + " " + user.LastName;
+
+                if (userName == comment.CommentUser)
+                {
+                    return View(comment);
+                }
+                else
+                {
+                    return RedirectToAction("PrivilegeError", "News");
+                }
+            }
+            else
+            {
+                return RedirectToAction("Login", "Account");
+            }            
         }
 
         // GET: Comments/Create
@@ -82,13 +120,35 @@ namespace FinalProject.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+
             Comment comment = db.Comments.Find(id);
+
             if (comment == null)
             {
                 return HttpNotFound();
             }
+
             ViewBag.ArticleID = new SelectList(db.Articles, "ID", "Title", comment.ArticleID);
-            return View(comment);
+
+            string strCurrentUserId = User.Identity.GetUserId();
+            ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(strCurrentUserId);
+
+            var userName = user.FirstName + " " + user.LastName;
+
+            if (User.IsInRole("Admin") || (userName == comment.CommentUser))
+            {
+                return View(comment);
+            }
+            else if (User.Identity.IsAuthenticated)
+            {
+                return RedirectToAction("PrivilegeError", "News");
+            }
+            else
+            {
+                return RedirectToAction("Login", "Account");
+            }
+
+
         }
 
         // POST: Comments/Edit/5
@@ -98,15 +158,35 @@ namespace FinalProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "ID,ArticleID,CommentTitle,CommentUser,Text, PublishDate")] Comment comment)
         {
-            if (ModelState.IsValid)
+
+            string strCurrentUserId = User.Identity.GetUserId();
+            ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(strCurrentUserId);
+
+            var userName = user.FirstName + " " + user.LastName;
+
+            if (User.IsInRole("Admin") || (userName == comment.CommentUser))
             {
-                comment.PublishDate = System.DateTime.Now;
-                db.Entry(comment).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    comment.PublishDate = System.DateTime.Now;
+                    db.Entry(comment).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.ArticleID = new SelectList(db.Articles, "ID", "Title", comment.ArticleID);
+                return View(comment);
             }
-            ViewBag.ArticleID = new SelectList(db.Articles, "ID", "Title", comment.ArticleID);
-            return View(comment);
+            else if (User.Identity.IsAuthenticated)
+            {
+                return RedirectToAction("PrivilegeError", "News");
+            }
+            else
+            {
+                return RedirectToAction("Login", "Account");
+            }
+
+
+            
         }
 
         // GET: Comments/Delete/5
@@ -121,7 +201,26 @@ namespace FinalProject.Controllers
             {
                 return HttpNotFound();
             }
-            return View(comment);
+
+            string strCurrentUserId = User.Identity.GetUserId();
+            ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(strCurrentUserId);
+
+            var userName = user.FirstName + " " + user.LastName;
+
+            if (User.IsInRole("Admin") || (userName == comment.CommentUser))
+            {                
+                return View(comment);
+            }
+            else if (User.Identity.IsAuthenticated)
+            {
+                return RedirectToAction("PrivilegeError", "News");
+            }
+            else
+            {
+                return RedirectToAction("Login", "Account");
+            }
+
+            
         }
 
         // POST: Comments/Delete/5
@@ -129,10 +228,29 @@ namespace FinalProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+
+            string strCurrentUserId = User.Identity.GetUserId();
+            ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(strCurrentUserId);
+
+            var userName = user.FirstName + " " + user.LastName;
             Comment comment = db.Comments.Find(id);
-            db.Comments.Remove(comment);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+
+            if (User.IsInRole("Admin") || (userName == comment.CommentUser))
+            {
+                db.Comments.Remove(comment);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else if (User.Identity.IsAuthenticated)
+            {
+                return RedirectToAction("PrivilegeError", "News");
+            }
+            else
+            {
+                return RedirectToAction("Login", "Account");
+            }
+
+            
         }
 
         protected override void Dispose(bool disposing)
